Start breaking down cli into modules from one giant file
