+++ cli/cli.py
@@ -1,87 +1,9 @@
 """CLI tool for interacting with the API."""
 
-from bike import Bike
-from wheel import FrontWheel, RearWheel
-import requests
+from bikefetcher import BikeFetcher
+from wheelfetcher import WheelFetcher
+from wheelinstaller import WheelInstaller
 import time
-import re
-from IPython import embed
-
-
-class BikeFetcher():
-    def get_bikes(self):
-        response = requests.get('http://localhost:8000/bikes/')
-        bike_info = response.json()
-        option = 1
-        bike_dict = {}
-        for bike_attributes in bike_info:
-            bike = Bike(bike_attributes)
-            bike_dict[option] = bike
-            option += 1
-        return bike_dict
-
-    def get_bike(self, bike_db_ref):
-        response = requests.get(f"http://localhost:8000/bikes/{bike_db_ref}")
-        bike_info = response.json()
-        bike = Bike(bike_info)
-        return bike
-
-
-class WheelFetcher():
-    def get_front_wheels(self):
-        response = requests.get('http://localhost:8000/wheels/front')
-        wheel_info = response.json()
-        front_wheels = []
-        for wheel_attributes in wheel_info:
-            wheel = FrontWheel(wheel_attributes)
-            front_wheels.append(wheel)
-        return front_wheels
-
-    def get_rear_wheels(self):
-        response = requests.get('http://localhost:8000/wheels/rear')
-        wheel_info = response.json()
-        rear_wheels = []
-        for wheel_attributes in wheel_info:
-            wheel = RearWheel(wheel_attributes)
-            rear_wheels.append(wheel)
-        return rear_wheels
-
-    def get_wheels(self):
-        fronts = self.get_front_wheels()
-        rears = self.get_rear_wheels()
-        all = fronts + rears
-        option = 1
-        wheel_dict = {}
-        for wheel in all:
-            wheel_dict[option] = wheel
-            option += 1
-        return wheel_dict
-
-
-class WheelInstaller():
-    """Install a wheel on a bike through the API."""
-
-    def install(self, wheel, bike):
-        """Install a wheel on the bike or handle error."""
-        bike_info = bike.__dict__.copy()
-        bike_info['brake_type'] = bike.brakes
-        location = 'f_wheel' if wheel.wheel_type == 'front' else 'r_wheel'
-        bike_info[location] = wheel.db_ref
-        response = requests.patch(
-            f"http://localhost:8000/bikes/{bike.db_ref}/",
-            bike_info,
-        )
-        if response.status_code == 200:
-            return f"{wheel.model.capitalize()} was installed!"
-        if response.status_code == 400:
-            if 'field_errors' in response.content.decode():
-                error = response.content.decode().split('[')[1]
-                result = re.findall('[A-z \s !]+', error)
-                return result[0]
-            else:
-                return "Looks like that wheel might not be compatible or one might already be on the bike."
-        else:
-            return "Something went wrong. Try again..."
 
 
 bikefetcher = BikeFetcher()
